complete the backbone and add the random agent
